fix(stellar): retry Horizon poll on transient errors, dedupe helpers, unstick batch
- Retry transaction lookup on 5xx/429 and transport-style errors like 404
- Share Stellar submit helpers between activation settlement and spend flows
- Remove batch skip so submitted rows without tx_hash reach failSettlement cleanup
