+++ lib/activation/stellar-settlement-submit.ts
@@ -7,79 +7,19 @@ import {
   xdr,
 } from '@stellar/stellar-sdk';
 import { privyWalletRawSignTransactionHash } from '@/lib/privy-server-rest';
+import { stellarWalletAddressSchema } from '@/lib/schemas/player';
 import { stellarUsdcAssetConfigSchema } from '@/lib/schemas/sponsored-activation';
+import {
+  classifyStellarHorizonSubmitError,
+  formatUsdcAmountForStellar,
+  loadAccountOrThrow,
+  readHorizonHttpErrorData,
+} from '@/lib/stellar/horizon-submit-helpers';
 import {
   createStellarSpendHorizonServer,
   getStellarSpendNetworkPassphrase,
   parseStellarSpendSponsorKeypair,
 } from '@/lib/spend/stellar-wallet-readiness-config';
-import { stellarWalletAddressSchema } from '@/lib/schemas/player';
-
-function formatUsdcAmountForStellar(amount: number): string {
-  const floored = Math.floor(amount * 1e7) / 1e7;
-  return floored.toFixed(7);
-}
-
-export function classifyStellarSettlementSubmitError(
-  raw: string,
-  resultCodes?: unknown
-): string {
-  const s = `${raw} ${JSON.stringify(resultCodes ?? {})}`.toLowerCase();
-  if (
-    s.includes('fetch') ||
-    s.includes('econn') ||
-    s.includes('timeout') ||
-    s.includes('socket') ||
-    s.includes('502') ||
-    s.includes('503') ||
-    s.includes('504')
-  ) {
-    return 'network_submit_failed';
-  }
-  if (
-    s.includes('underfunded') ||
-    s.includes('op_underfunded') ||
-    s.includes('op_low_reserve') ||
-    s.includes('insufficient balance')
-  ) {
-    return 'insufficient_usdc_or_reserve';
-  }
-  return 'stellar_submit_failed';
-}
-
-function readHorizonHttpErrorData(e: unknown): {
-  data: Record<string, unknown> | null;
-  resultCodes: unknown;
-} {
-  const raw = (e as { response?: { data?: unknown } })?.response?.data;
-  if (!raw || typeof raw !== 'object') {
-    return { data: null, resultCodes: null };
-  }
-  const data = raw as Record<string, unknown>;
-  const extras = data.extras;
-  if (!extras || typeof extras !== 'object') {
-    return { data, resultCodes: null };
-  }
-  return {
-    data,
-    resultCodes: (extras as { result_codes?: unknown }).result_codes ?? null,
-  };
-}
-
-async function loadAccountOrThrow(
-  server: Horizon.Server,
-  publicKey: string
-): Promise<Horizon.AccountResponse> {
-  try {
-    return await server.loadAccount(publicKey);
-  } catch (e: unknown) {
-    const status = (e as { response?: { status?: number } })?.response?.status;
-    if (status === 404) {
-      throw new Error('source_account_missing');
-    }
-    throw e;
-  }
-}
 
 export type StellarSettlementAssetConfig = {
   asset_code: string;
@@ -148,7 +88,7 @@ export async function submitStellarActivationSettlementFromCampaign(input: {
     const { resultCodes } = readHorizonHttpErrorData(e);
     return {
       ok: false,
-      reason: classifyStellarSettlementSubmitError(msg, resultCodes),
+      reason: classifyStellarHorizonSubmitError(msg, resultCodes),
       internalMessage: msg.slice(0, 500),
     };
   }
@@ -191,7 +131,7 @@ export async function submitStellarActivationSettlementFromCampaign(input: {
     const msg = e instanceof Error ? e.message : String(e);
     return {
       ok: false,
-      reason: classifyStellarSettlementSubmitError(msg),
+      reason: classifyStellarHorizonSubmitError(msg),
       internalMessage: msg.slice(0, 500),
     };
   }
@@ -211,7 +151,7 @@ export async function submitStellarActivationSettlementFromCampaign(input: {
     const msg = e instanceof Error ? e.message : String(e);
     return {
       ok: false,
-      reason: classifyStellarSettlementSubmitError(msg),
+      reason: classifyStellarHorizonSubmitError(msg),
       internalMessage: msg.slice(0, 500),
     };
   }
@@ -232,7 +172,7 @@ export async function submitStellarActivationSettlementFromCampaign(input: {
     const { data, resultCodes } = readHorizonHttpErrorData(e);
     return {
       ok: false,
-      reason: classifyStellarSettlementSubmitError(msg, resultCodes),
+      reason: classifyStellarHorizonSubmitError(msg, resultCodes),
       internalMessage: JSON.stringify({
         message: msg.slice(0, 400),
         ...(data?.title ? { horizon_title: data.title } : {}),
