+++ lib/spend/stellar-spend-payment-submit.ts
@@ -8,60 +8,18 @@ import {
 } from '@stellar/stellar-sdk';
 import { privyWalletRawSignTransactionHash } from '@/lib/privy-server-rest';
 import {
+  classifyStellarHorizonSubmitError,
+  formatUsdcAmountForStellar,
+  loadAccountOrThrow,
+  readHorizonHttpErrorData,
+} from '@/lib/stellar/horizon-submit-helpers';
+import {
   createStellarSpendHorizonServer,
   getStellarSpendNetworkPassphrase,
   getStellarSpendUsdcAssetCode,
   getStellarSpendUsdcIssuer,
   parseStellarSpendSponsorKeypair,
 } from '@/lib/spend/stellar-wallet-readiness-config';
-
-function formatUsdcAmountForStellar(amount: number): string {
-  const floored = Math.floor(amount * 1e7) / 1e7;
-  return floored.toFixed(7);
-}
-
-function classifySubmitMessage(raw: string, resultCodes?: unknown): string {
-  const s = `${raw} ${JSON.stringify(resultCodes ?? {})}`.toLowerCase();
-  if (
-    s.includes('fetch') ||
-    s.includes('econn') ||
-    s.includes('timeout') ||
-    s.includes('socket') ||
-    s.includes('502') ||
-    s.includes('503') ||
-    s.includes('504')
-  ) {
-    return 'network_submit_failed';
-  }
-  if (
-    s.includes('underfunded') ||
-    s.includes('op_underfunded') ||
-    s.includes('op_low_reserve') ||
-    s.includes('insufficient balance')
-  ) {
-    return 'insufficient_usdc_or_reserve';
-  }
-  return 'stellar_submit_failed';
-}
-
-function readHorizonHttpErrorData(e: unknown): {
-  data: Record<string, unknown> | null;
-  resultCodes: unknown;
-} {
-  const raw = (e as { response?: { data?: unknown } })?.response?.data;
-  if (!raw || typeof raw !== 'object') {
-    return { data: null, resultCodes: null };
-  }
-  const data = raw as Record<string, unknown>;
-  const extras = data.extras;
-  if (!extras || typeof extras !== 'object') {
-    return { data, resultCodes: null };
-  }
-  return {
-    data,
-    resultCodes: (extras as { result_codes?: unknown }).result_codes ?? null,
-  };
-}
 
 function formatHorizonSubmitInternalMessage(
   msg: string,
@@ -76,21 +34,6 @@ function formatHorizonSubmitInternalMessage(
     ...(detail ? { horizon_detail: detail.slice(0, 800) } : {}),
     ...(resultCodes ? { horizon_result_codes: resultCodes } : {}),
   }).slice(0, 1200);
-}
-
-async function loadAccountOrThrow(
-  server: Horizon.Server,
-  publicKey: string
-): Promise<Horizon.AccountResponse> {
-  try {
-    return await server.loadAccount(publicKey);
-  } catch (e: unknown) {
-    const status = (e as { response?: { status?: number } })?.response?.status;
-    if (status === 404) {
-      throw new Error('source_account_missing');
-    }
-    throw e;
-  }
 }
 
 /**
@@ -134,7 +77,7 @@ export async function submitSponsoredStellarUsdcPaymentFromUser(input: {
     const { resultCodes } = readHorizonHttpErrorData(e);
     return {
       ok: false,
-      reason: classifySubmitMessage(msg, resultCodes),
+      reason: classifyStellarHorizonSubmitError(msg, resultCodes),
       internalMessage: msg.slice(0, 500),
     };
   }
@@ -177,7 +120,7 @@ export async function submitSponsoredStellarUsdcPaymentFromUser(input: {
     const msg = e instanceof Error ? e.message : String(e);
     return {
       ok: false,
-      reason: classifySubmitMessage(msg),
+      reason: classifyStellarHorizonSubmitError(msg),
       internalMessage: msg.slice(0, 500),
     };
   }
@@ -197,7 +140,7 @@ export async function submitSponsoredStellarUsdcPaymentFromUser(input: {
     const msg = e instanceof Error ? e.message : String(e);
     return {
       ok: false,
-      reason: classifySubmitMessage(msg),
+      reason: classifyStellarHorizonSubmitError(msg),
       internalMessage: msg.slice(0, 500),
     };
   }
@@ -218,7 +161,7 @@ export async function submitSponsoredStellarUsdcPaymentFromUser(input: {
     const { data, resultCodes } = readHorizonHttpErrorData(e);
     return {
       ok: false,
-      reason: classifySubmitMessage(msg, resultCodes),
+      reason: classifyStellarHorizonSubmitError(msg, resultCodes),
       internalMessage: formatHorizonSubmitInternalMessage(
         msg,
         data,
